+++ tests/experimental/test_pipe.py
@@ -6,6 +6,7 @@
 
 import asyncio
 import typing
+from contextlib import AsyncExitStack, aclosing, closing
 
 from frequenz.channels import Broadcast, Receiver
 from frequenz.channels.experimental import Pipe
@@ -55,3 +56,27 @@ async def test_pipe() -> None:
     await sender_chan2.send(5)
     assert await receive_timeout(receiver_chan1) is Timeout
     assert await receive_timeout(receiver_chan2) == 5
+
+
+async def test_pipe_with_async_exit_stack() -> None:
+    """Test how pipe should work with async exit stack."""
+    # Channels, receivers and Pipe are in AsyncExitStack
+    # to close and stop them at the end.
+    async with AsyncExitStack() as stack:
+        source_channel = await stack.enter_async_context(
+            aclosing(Broadcast[int](name="source channel"))
+        )
+        source_receiver = stack.enter_context(closing(source_channel.new_receiver()))
+
+        forwarding_channel = await stack.enter_async_context(
+            aclosing(Broadcast[int](name="forwarding channel"))
+        )
+        await stack.enter_async_context(
+            Pipe(source_receiver, forwarding_channel.new_sender())
+        )
+
+        receiver = stack.enter_context(closing(forwarding_channel.new_receiver()))
+
+        source_sender = source_channel.new_sender()
+        await source_sender.send(10)
+        assert await receiver.receive() == 10